Changes to be committed: modified:   src/common/column/boolean.column.ts

+++ src/common/column/boolean.column.ts
@@ -32,7 +32,7 @@ export function BooleanColumn(
       default: defaultValue,
       name,
       transformer: new BooleanColumnTransformer(),
-      type: 'int',
+      type: 'smallint',
       width: 1,
     })(object, propertyName);
   };
